+++ package.js
@@ -2,7 +2,7 @@ Package.describe({
   name: "aldeed:autoform",
   summary: "Easily create forms with automatic insert and update, and automatic reactive validation.",
   git: "https://github.com/aldeed/meteor-autoform.git",
-  version: "2.0.2"
+  version: "3.0.0"
 });
 
 Package.on_use(function(api) {